# services/api/src/main/java/com/taronote/content/web/CreateNoteRequest.java
package com.taronote.content.web;

import jakarta.validation.constraints.NotBlank;
import java.util.List;

public record CreateNoteRequest(
        @NotBlank String title,
        String content,
        List<String> images
) {
}