# services/api/src/main/java/com/taronote/identity/web/RegisterRequest.java
package com.taronote.identity.web;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record RegisterRequest(
        @NotBlank String username,
        @Email @NotBlank String email,
        @NotBlank String password
) {
}